Upgrade Ruby on Rails to 5.0 (#2095)
Migration to newer **rails** gem version: `5.0.7`.

We suggest to following tasks in the topic after this will be merged:
- Update grape and relative gems.
- Update to ruby **2.6**.
- Update to rails **5.1**.

+++ db/migrate/20180529125011_replace_id_to_code.rb
@@ -1,4 +1,4 @@
-class ReplaceIdToCode < ActiveRecord::Migration
+class ReplaceIdToCode < ActiveRecord::Migration[4.2]
   def change
     ActiveRecord::Base.transaction do
       execute %[ALTER TABLE `currencies` CHANGE `code` `code` VARCHAR(10) CHARACTER SET utf8 COLLATE utf8_unicode_ci NOT NULL;]
@@ -25,7 +25,11 @@ class ReplaceIdToCode < ActiveRecord::Migration
 
       execute %[UPDATE `currencies` SET `id` = `code`;]
       execute %[ALTER TABLE `currencies` DROP `code`;]
-      remove_index :currencies, column: %i[enabled code]
+
+      if index_exists?(:currencies, %i[enabled code])
+        remove_index :currencies, column: %i[enabled code]
+      end
+
       add_index :currencies, [:enabled]
     end
   end
